feat: add ability to cancel own bookings

+++ pages/api/booking/[id].ts
@@ -1,8 +1,8 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
 import { getSession } from 'next-auth/client';
-import { isModerator } from '../../../../lib/authorization';
-import { sendCancelationEmail } from '../../../../lib/email/cancel';
-import prisma from '../../../../lib/prisma';
+import { isModerator } from '../../../lib/authorization';
+import { sendCancelationEmail } from '../../../lib/email/cancel';
+import prisma from '../../../lib/prisma';
 
 export default async (req: NextApiRequest, res: NextApiResponse) => {
     if (req.method !== 'DELETE') {
@@ -17,11 +17,6 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
         return;
     }
 
-    if (!isModerator(session)) {
-        res.status(401).json({ result: 'error' });
-        return;
-    }
-
     const id = parseInt(req.query.id.toString(), 10);
 
     if (isNaN(id)) {
@@ -29,11 +24,32 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
         return;
     }
 
-    const booking = await prisma.booking.delete({
+    let booking = await prisma.booking.findUnique({
         where: {
             id,
         }
     });
+
+    if (!booking) {
+        res.status(400).json({ result: 'error', message: 'Booking does not exist' });
+        return;
+    }
+
+    if (new Date(booking.date) < new Date() || booking.result !== 'unknown') {
+        res.status(400).json({ result: 'error', message: 'Can not delete old or processed bookings' });
+        return;
+    }
+
+    if (isModerator(session) || (booking.email && session.user?.email === booking.email)) {
+        booking = await prisma.booking.delete({
+            where: {
+                id,
+            }
+        });
+    } else {
+        res.status(401).json({ result: 'error', message: 'No permission' });
+        return;
+    }
 
     try {
         await sendCancelationEmail(booking);
